requirements.txt
Django==2.2
django-filter==2.1.0
django-graphql-jwt==0.2.1
graphene-django==2.2.0
psycopg2
django-cors-headers
graphene-file-upload
pandas
xlrd
celery
py-graphql-client
pyodbc
graphqlclient
redis
openpyxl
